+++ eyes.sdk.core/src/main/java/com/applitools/eyes/MobileDeviceInfo.java
@@ -3,6 +3,7 @@ package com.applitools.eyes;
 public class MobileDeviceInfo {
     private double pixelRatio;
     private Region viewportRect;
+    private String[] aliases;
 
     public double getPixelRatio() {
         return pixelRatio;
@@ -18,5 +19,13 @@ public class MobileDeviceInfo {
 
     public void setViewportRect(Region viewportRect) {
         this.viewportRect = viewportRect;
+    }
+
+    public String[] getAliases() {
+        return aliases;
+    }
+
+    public void setAliases(String[] aliases) {
+        this.aliases = aliases;
     }
 }